requirements.txt
fastapi
uvicorn
frida-tools
google-generativeai
python-dotenv
apktool-py
scapy
